requirements.txt
pyyaml
ollama
openai
python-dotenv

autogen-core
autogen-ext[docker]
beautifulsoup4
aiofiles
requests
mammoth
markdownify
numpy
python-pptx
pandas
openpyxl
pdfminer.six
puremagic
pydub
youtube-transcript-api
SpeechRecognition
pathvalidate
playwright
pydantic<3.0.0,>=2.0.0
plotly

pydantic-ai